add 보드게임 상세 페이지 컴포넌트 분리, 보드게임 좋아요 누른 사람들 모달 내 컴포넌트
