+++ src/components/boardgame/BoardgameLikedPerson.css
@@ -1,0 +1,7 @@
+.div-boardgame-liked {
+  display: flex;
+  justify-content: space-evenly;
+  align-items: center;
+  padding: 5%;
+}
+
