+++ src/components/boardgame/BoardgameTagInfo.css
@@ -1,0 +1,43 @@
+.div-boardgame-best-tags {
+  background-image: url('/public/img/F3_best_tags_background.png'); /* 배경 이미지 경로 */
+  background-size: 100vw auto; /* 너비를 100vw로 고정, 비율 유지 */
+  background-repeat: no-repeat; /* 이미지를 반복하지 않음 */
+  background-position: top;  /* 이미지를 상단 정렬 */
+  height: auto;                 /* 높이는 비율 유지 */
+  max-width: 800px;                 
+  max-height: 100%;             /* 높이가 화면 높이를 넘지 않도록 제한 */
+  margin: var(--basic-margin) var(--basic-margin) 0 var(--basic-margin);
+  padding: 4%;
+  border-radius: var(--basic-border-radius);
+  color: white;
+}
+
+
+.div-title-tag{
+  margin: 2% 2% 4% 2%;
+  padding: 1% 4% 1% 3%;
+  border-radius: var(--basic-border-radius);
+  background-color: var(--basic-gray-color);
+  display: flex;
+  align-items: left;
+  text-align: left;
+  color: black;
+  font-size: xx-large;
+  border : 2px solid black;
+}
+
+.div-selected-tag{
+  margin: 1% var(--basic-margin) 1% var(--basic-margin);
+  padding: 1% 4% 1% 3%;
+  border-radius: var(--basic-border-radius);
+  background-color: var(--basic-gray-color);
+  display: flex;
+  align-items: left;
+  text-align: left;
+  color: black;
+  font-size: x-large;
+  border : 2px solid black;
+}
+.div-boardgame-ratings{
+  flex-grow: 1;
+}
