+++ src/pages/boardgame/BoardGameDetail.css
@@ -2,121 +2,6 @@
   background-color: var(--basic-white-color);
   display: flex;
   flex-direction: column; /* 세로 방향 정렬 */
-  /* height: inherit; */
-}
-
-.div-boardgame-image {
-  height : 40%;
-  background-color: var(--basic-black-color);
-}
-
-.div-boardgame-title-info {
-  height : 30%;
-  background-color: var(--basic-black-color);
-  color: white;
-}
-
-.div-boardgame-title-info > div {
-  text-align: left;
-}
-
-.div-boardgame-title-info > div > span {
-  font-size : x-large;
-  font-weight: bold;
-}
-
-.div-cast-info {
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  padding: 3% 3% ;
-}
-
-.div-cast-info > div {
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-}
-
-.div-cast-info > div > div{
-  text-align: center;
-  font-size: xxx-large;
-  font-weight: bold;
-}
-
-.basic-boardgame-comment {
-  height : 30%;
-}
-
-.div-boardgame-more-info {
-  table-layout: fixed;
-  width: 100%;
-  border-collapse: collapse;
-  background-color: #e7e7e9;
-}
-
-.div-boardgame-more-info > tr {
-  height: 6vh;
-}
-
-.div-boardgame-more-info > tr > td {
-  border : 1px solid #c6c6c7;
-  font-size: xx-large;
-  font-weight: bold;
-}
-
-.td-weight-info {
-  width:100%;
-  padding-left: 8%;
-  padding-right: 8%;
-}
-
-.table-inner-info{
-  font-size: medium;
-}
-
-.div-boardgame-best-tags {
-  background-image: url('/public/img/F3_best_tags_background.png'); /* 배경 이미지 경로 */
-  background-size: 100vw auto; /* 너비를 100vw로 고정, 비율 유지 */
-  background-repeat: no-repeat; /* 이미지를 반복하지 않음 */
-  background-position: top;  /* 이미지를 상단 정렬 */
-  height: auto;                 /* 높이는 비율 유지 */
-  max-width: 800px;                 
-  max-height: 100%;             /* 높이가 화면 높이를 넘지 않도록 제한 */
-  margin: var(--basic-margin) var(--basic-margin) 0 var(--basic-margin);
-  padding: 4%;
-  border-radius: var(--basic-border-radius);
-  color: white;
-}
-
-.div-boardgame-ratings{
-  flex-grow: 1;
-}
-
-.div-title-tag{
-  margin: 2% 2% 4% 2%;
-  padding: 1% 4% 1% 3%;
-  border-radius: var(--basic-border-radius);
-  background-color: var(--basic-gray-color);
-  display: flex;
-  align-items: left;
-  text-align: left;
-  color: black;
-  font-size: xx-large;
-  border : 2px solid black;
-}
-
-.div-selected-tag{
-  margin: 1% var(--basic-margin) 1% var(--basic-margin);
-  padding: 1% 4% 1% 3%;
-  border-radius: var(--basic-border-radius);
-  background-color: var(--basic-gray-color);
-  display: flex;
-  align-items: left;
-  text-align: left;
-  color: black;
-  font-size: x-large;
-  border : 2px solid black;
 }
 
 .create-rating-btn{
